settings: pull settings out into its own file
